33.0
Due to an influx in malware on AUR changes are being made sorry for any inconvenience. We choose safety first.
With that said installation of Yay and all packages that previously were installed with Yay have been removed.
Also switched OS drive format to BTRFS due to more security features.